requirements.txt
streamlit
torch
torchvision
Pillow
opencv-python
numpy
tensorflow
keras
transformers
sentencepiece
pycocoevalcap